refactor(querybuilder): share the JSON text literal helpers
JSONEscapable, JSONTextRuns and JSONComparisonLiterals carry nothing
body-specific, so they move next to ClickHouseLikePatternLiteral for
any signal filtering over JSON-encoded text in a string column.
